Move all services to new service package, split AusleihServiceController into ItemController and PersonController, adjust scopes accordingly
