+++ ausleih-service/src/main/java/hhu/ausleihservice/web/service/ItemService.java
@@ -1,4 +1,4 @@
-package hhu.ausleihservice.web;
+package hhu.ausleihservice.web.service;
 
 import hhu.ausleihservice.dataaccess.ItemRepository;
 import hhu.ausleihservice.databasemodel.Item;
@@ -17,12 +17,12 @@ public class ItemService {
 	private ItemAvailabilityService itemAvailabilityService;
 
 
-	ItemService(ItemRepository itemRep, ItemAvailabilityService itemAvailabilityService) {
+	public ItemService(ItemRepository itemRep, ItemAvailabilityService itemAvailabilityService) {
 		this.items = itemRep;
 		this.itemAvailabilityService = itemAvailabilityService;
 	}
 
-	Item findById(long id) {
+	public Item findById(long id) {
 		Optional<Item> item = items.findById(id);
 		if (!item.isPresent()) {
 			throw new ItemNichtVorhanden();
