+++ ausleih-service/src/main/resources/templates/artikelListe.html
@@ -18,7 +18,6 @@
 		<th>Tagessatz</th>
 		<th>Kaution</th>
 		<th>Standort</th>
-		<th th:if="${user.isAdmin()}"> </th>
 	</tr>
 	<th:block th:each="artikel : ${artikelListe}">
 		<tr>
